python/orchestrator.py
"""
Orchestrator module for WhatsApp Agent
Initializes Agno-like workflow orchestration for multi-step LLM tasks
"""

import logging
import os
from dotenv import load_dotenv

logger = logging.getLogger(__name__)

load_dotenv()

# ============================================================================
# ORCHESTRATOR STATE
# ============================================================================

orchestrator_state = {
    'initialized': False,
    'agents': {},
    'workflows': {},
    'message_history': [],
}

# ============================================================================
# AGENT DEFINITIONS (Agno-like structure)
# ============================================================================

class Agent:
    """
    Simple agent wrapper for Agno-like interface
    """
    def __init__(self, name, description=''):
        self.name = name
        self.description = description
        self.tools = []
        self.history = []
        
    def run(self, task):
        """
        Run agent with a task
        """
        logger.info(f'Agent {self.name} running: {task}')
        return {'status': 'success', 'output': 'Placeholder agent output'}

# ============================================================================
# WORKFLOW DEFINITIONS
# ============================================================================

class Workflow:
    """
    Simple workflow for orchestrating multi-step tasks
    """
    def __init__(self, name, steps):
        self.name = name
        self.steps = steps
        self.state = {}
        
    def execute(self, input_data):
        """
        Execute workflow steps sequentially
        """
        logger.info(f'Executing workflow: {self.name}')
        current_input = input_data
        
        for step in self.steps:
            logger.debug(f'  Step: {step["name"]}')
            # In Phase 2+, this will call actual agents/models
            current_input = {'status': 'success', 'data': current_input}
        
        return current_input

# ============================================================================
# INITIALIZE ORCHESTRATOR
# ============================================================================

def initialize_orchestrator():
    """
    Initialize Agno-like orchestration framework
    Setup agents and workflows for summarization, mention detection, and alerting
    """
    try:
        logger.info('🔧 Initializing orchestrator...')

        # Create agents
        summarizer_agent = Agent(
            name='SummarizerAgent',
            description='Summarizes long text using DistilBART'
        )
        
        mention_detection_agent = Agent(
            name='MentionDetectorAgent',
            description='Detects mentions and extracts actionable tasks'
        )
        
        alert_generation_agent = Agent(
            name='AlertGeneratorAgent',
            description='Generates alerts for user-actionable items'
        )

        orchestrator_state['agents'] = {
            'summarizer': summarizer_agent,
            'mention_detector': mention_detection_agent,
            'alert_generator': alert_generation_agent,
        }

        # Create workflows
        summarization_workflow = Workflow(
            name='summarization',
            steps=[
                {'name': 'load_text', 'agent': 'summarizer'},
                {'name': 'preprocess', 'agent': 'summarizer'},
                {'name': 'generate_summary', 'agent': 'summarizer'},
                {'name': 'post_process', 'agent': 'summarizer'},
            ]
        )

        mention_workflow = Workflow(
            name='mention_detection',
            steps=[
                {'name': 'extract_mentions', 'agent': 'mention_detector'},
                {'name': 'extract_keywords', 'agent': 'mention_detector'},
                {'name': 'classify_urgency', 'agent': 'mention_detector'},
            ]
        )

        alert_workflow = Workflow(
            name='alert_generation',
            steps=[
                {'name': 'detect_actionable_items', 'agent': 'alert_generator'},
                {'name': 'format_alert', 'agent': 'alert_generator'},
                {'name': 'determine_priority', 'agent': 'alert_generator'},
            ]
        )

        orchestrator_state['workflows'] = {
            'summarization': summarization_workflow,
            'mention_detection': mention_workflow,
            'alert_generation': alert_workflow,
        }

        orchestrator_state['initialized'] = True
        logger.info('✅ Orchestrator initialized successfully')
        logger.info(f'  Agents: {", ".join(orchestrator_state["agents"].keys())}')
        logger.info(f'  Workflows: {", ".join(orchestrator_state["workflows"].keys())}')

    except Exception as e:
        logger.error(f'❌ Failed to initialize orchestrator: {str(e)}')
        raise

# ============================================================================
# WORKFLOW EXECUTION HELPERS
# ============================================================================

def run_workflow(workflow_name, input_data):
    """
    Execute a workflow by name
    """
    if not orchestrator_state['initialized']:
        raise RuntimeError('Orchestrator not initialized')
    
    if workflow_name not in orchestrator_state['workflows']:
        raise ValueError(f'Workflow {workflow_name} not found')
    
    workflow = orchestrator_state['workflows'][workflow_name]
    return workflow.execute(input_data)

def get_agent(agent_name):
    """
    Get an agent by name
    """
    return orchestrator_state['agents'].get(agent_name)

def get_orchestrator_state():
    """
    Get current orchestrator state
    """
    return {
        'initialized': orchestrator_state['initialized'],
        'agent_count': len(orchestrator_state['agents']),
        'workflow_count': len(orchestrator_state['workflows']),
        'agents': list(orchestrator_state['agents'].keys()),
        'workflows': list(orchestrator_state['workflows'].keys()),
    }
